Windows: Fixes for 64-bit build.

+++ Programs/serial_windows.c
@@ -489,7 +489,7 @@ serialEnsureFileDescriptor (SerialDevice *serial) {
   if ((serial->fileDescriptor = cygwin_attach_handle_to_fd("serialdevice", -1, serial->package.fileHandle, TRUE, GENERIC_READ|GENERIC_WRITE)) >= 0) return 1;
   logSystemError("cygwin_attach_handle_to_fd");
 #else /* __CYGWIN__ */
-  if ((serial->fileDescriptor = _open_osfhandle((long)serial->package.fileHandle, O_RDWR)) >= 0) return 1;
+  if ((serial->fileDescriptor = _open_osfhandle((intptr_t)serial->package.fileHandle, O_RDWR)) >= 0) return 1;
   logSystemError("open_osfhandle");
 #endif /* __CYGWIN__ */
